refactor: address PR review feedback
- Extract high scores persistence into scores.lua module
- Extract NameEntry and HighScoresView into highscores.lua module
- Use DEFER for PHYSFS_close in Spit/Slurp
- Add ReadJson helper to json_alc.h, use it in config.cc, packer.cc,
  and lua_json.cc to reduce yyjson boilerplate
- Remove obvious comment from lua_json.h
