+++ src/json_alc.h
@@ -34,6 +34,20 @@ inline yyjson_alc MakeYyjsonAlc(ArenaAllocator* arena) {
   return alc;
 }
 
+// Parse a JSON string using the given arena allocator. Returns the root value
+// on success, or nullptr on failure (with err populated).
+inline yyjson_doc* ReadJson(ArenaAllocator* arena, const char* data, size_t len,
+                            yyjson_read_err* err) {
+  yyjson_alc alc = MakeYyjsonAlc(arena);
+  return yyjson_read_opts(const_cast<char*>(data), len, YYJSON_READ_NOFLAG,
+                          &alc, err);
+}
+
+inline yyjson_doc* ReadJson(ArenaAllocator* arena, std::string_view str,
+                            yyjson_read_err* err) {
+  return ReadJson(arena, str.data(), str.size(), err);
+}
+
 // View the string payload of a yyjson value as a std::string_view.
 // The view is valid for as long as the owning yyjson_doc (and its arena)
 // lives.
