+++ src/lua_json.cc
@@ -133,10 +133,8 @@ const struct LuaApiFunction kJsonLib[] = {
        size_t len = 0;
        const char* str = luaL_checklstring(state, 1, &len);
        auto* arena = Registry<ArenaAllocator>::Retrieve(state);
-       yyjson_alc alc = MakeYyjsonAlc(arena);
        yyjson_read_err err{};
-       yyjson_doc* doc = yyjson_read_opts(const_cast<char*>(str), len,
-                                          YYJSON_READ_NOFLAG, &alc, &err);
+       yyjson_doc* doc = ReadJson(arena, str, len, &err);
        if (doc == nullptr) {
          lua_pushlstring(state, err.msg, err.msg ? strlen(err.msg) : 0);
          lua_pushnil(state);
@@ -154,7 +152,8 @@ const struct LuaApiFunction kJsonLib[] = {
      [](lua_State* state) {
        auto* arena = Registry<ArenaAllocator>::Retrieve(state);
        yyjson_alc alc = MakeYyjsonAlc(arena);
-       yyjson_mut_doc* doc = yyjson_mut_doc_new(&alc);
+       yyjson_mut_doc* doc =
+           yyjson_mut_doc_new(&alc);  // mutable doc for encode
        auto result = LuaToJsonValue(state, 1, doc);
        if (result.is_error()) {
          auto msg = result.error().message();
